Remove X digest long-term memory
